feat : 도메인 수정
- UketEventRound로 ticketingDate 이동
- Payment로 결제 정보 이동
- Payment를 Organization이 아닌 Event와 연관관계를 맺도록 수정
- nowTicketing validation에 ticketing status를 이용하지 않도록 수정
