+++ src/main/kotlin/uket/domain/payment/entity/Payment.kt
@@ -1,5 +1,6 @@
 package uket.domain.payment.entity
 
+import jakarta.persistence.Column
 import jakarta.persistence.Entity
 import jakarta.persistence.GeneratedValue
 import jakarta.persistence.GenerationType
@@ -13,7 +14,18 @@ class Payment(
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     val id: Long = 0L,
-    val organizationId: Long,
-    val accountNo: String,
-    val depositLink: String,
+
+    @Column(name = "uket_event_id")
+    val uketEventId: Long,
+
+    @Column(name = "ticket_price")
+    val ticketPrice: Long,
+    @Column(name = "bank_code")
+    val bankCode: String,
+    @Column(name = "account_number")
+    val accountNumber: String,
+    @Column(name = "depositor_name")
+    val depositorName: String,
+    @Column(name = "deposit_url")
+    val depositUrl: String,
 ) : BaseTimeEntity()
